backend/requirements.txt
django==4.2.*
djangorestframework==3.14.*
mongoengine==0.27.*
pymongo==4.3.3.*
python-dotenv==1.0.*
django-cors-headers==4.3.*
gunicorn==21.2.*
bcrypt==4.0.*
pyjwt==2.8.*